Add InputSymbolMap attribute to ConfuserProject to allow the map for the future runs
Sometimes it is necessary to provide a patch for obfuscated software. Having different obfuscation applied to patched version comparing with the released one effectively means the patched version cannot be deployed without replacing every obfuscated assembly.

This change makes it possible to provide a symbol map from a previous obfuscation and preserves the renames specified there.